requirements.txt
Django~=4.2
djangorestframework~=3.14
opencv-python~=4.8.0
numpy~=1.24
tensorflow~=2.14.0
